requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==5.0.1
django-environ==0.11.2
django-phonenumber-field==7.3.0
pillow==10.2.0
psycopg2==2.9.9
sqlparse==0.4.4
tzdata==2023.4
